isolate out fused_all_reduce
